Cypress: More cy test structure

+++ cypress/plugins/index.js
@@ -1,4 +1,4 @@
-const fs = require("fs");
+const fs = require("fs-extra");
 
 /// <reference types="cypress" />
 // ***********************************************************
@@ -19,16 +19,30 @@ const fs = require("fs");
  */
 // eslint-disable-next-line no-unused-vars
 module.exports = (on, config) => {
-    on("task", {
-        resetSetup() {
-            fs.rmSync("settings/users", { recursive: true, force: true });
-            fs.rmSync("settings/settings.yaml");
-            fs.copyFileSync(
-                "cypress/fixtures/01_setup/default-settings.yaml",
-                "settings/settings.yaml"
-            );
+  on("task", {
+    resetSetup() {
+      const users = "settings/users";
+      const settings = "settings/settings.yaml";
+      // of course files need to exist in order to perform a delete
+      if (fs.existsSync(settings) && fs.existsSync(users)) {
+        fs.rmSync(users, { recursive: true, force: true });
+        fs.rmSync(settings);
+        fs.copyFileSync(
+          "cypress/fixtures/01_setup/default-settings.yaml",
+          "settings/settings.yaml"
+        );
+      }
 
-            return null;
-        },
-    });
+      return null;
+    },
+    prepopulateSetup() {
+      const settings = "settings";
+      const settingsFixture =
+        "cypress/fixtures/01_setup/prepulate_settings_seed/settings";
+      // of course files need to exist in order to perform a delete
+      fs.copySync(settingsFixture, settings);
+
+      return null;
+    },
+  });
 };
